Former-commit-id: 51e50982ca182b29305cbef3cd4752ba57a56bf1

+++ rs-dam/indexer/indexer-domain/src/main/java/fr/cnes/regards/modules/indexer/domain/reminder/SearchAfterReminder.java
@@ -39,7 +39,7 @@ public class SearchAfterReminder extends AbstractReminder {
     private String searchAfterSortValues;
 
     // For information (serialized into Elasticsearch)
-    private final int nextOffset;
+    private final long nextOffset;
 
     // For information (serialized into Elasticsearch)
     private final int nextPageSize;
